casuarina: Switch majority of blobs to FP3
Since casuarina is pretty much a generic QRD with zero changes to BSP
from Vsmart, update most of the blobs to FP3 6.A.031.5.

While majority of the blobs are on FP3 6.A.031.5, it's safe to say that vendor
security patch should be updated to match with common blobs, instead of casuarina
old security patch level.

Also, extract more dependencies to pass ELF checks.

Change-Id: I65051fb54ac118e47a1944bbfe14eb97431e9699